Extract authentication view to its own component
